MNT Add light type annotations to skops.io (#219)
